factor

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -5,9 +5,9 @@
 */
 int main(void)
 {
-long int n;
+int n;
 n = 612852475143;
-long int div = 2, ans = 0, maxFact;
+int div = 2, ans = 0, maxFact;
 
 while (n != 0)
 {
@@ -19,7 +19,7 @@ maxFact = n;
 n = n / div;
 if (n == 1)
 {
-printf("%ld\n", maxFact);
+printf("%d\n", maxFact);
 ans = 1;
 break;
 }
